+++ drizzle/0013_lucky_multiple_man.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `song` ADD `sections` text DEFAULT '[]' NOT NULL;